Add example of usage with an explicit instance of Joined
Instead of providing it implicitly, we test that we can plug our own
Joined instances. This is non-typesafe, but as the user is explicitly
putting their values we may assume that's expected behavior at this
point. However, a better abstraction could be put into place (e.g.,
having our own Joined wrapper that disallows the wildcard type
behaviors).